changed tipping view controllers to be a subclass
This shows off
OOP, polymorphism, inheritance and all the other things that hiring managers ask people about.
Why did I do this, to keep my code DRY (don’t repeat yoruself) and to make a more ridgid structure for the subclasses to follow.
Its good to do this because I could divy up the work to other programmers now. Here are some helpful constraints that I must design in now and if I need to change the superclass for some reason all the other subclasses should benefit from that change
THe Unlockable Tab Bar controller needs to be worked on but that will be the next commit